requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
pillow==9.4.0
typing-extensions==4.5.0
tqdm==4.62.3
urllib3==1.26.15

